% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curves.R
\name{as_survival_curves}
\alias{as_survival_curves}
\alias{as_survival_curves.data.frame}
\title{Convert to a \code{survival_curves} object}
\usage{
as_survival_curves(x, ...)

\method{as_survival_curves}{data.frame}(x, time = "time", id = "id", survival = "survival", ...)
}
\arguments{
\item{x}{An object of the appropriate class.}

\item{...}{Additional arguments to pass to \code{\link[=survival_curves]{survival_curves()}}.}

\item{time}{The name of the column containing the times a which survival
probabilities were computed.}

\item{id}{The name of the column containing an indentifier for each individual.}

\item{survival}{The name of the column containing survival probabilities.}
}
\value{
A \code{\link{survival_curves}} object.
}
\description{
\code{\link[=as_survival_curves]{as_survival_curves()}} is a generic function for converting objects to
the \code{survival_curves} class. There is currently only a method for a
\code{data.frame}.
}
\examples{
# Create a data frame storing survival curves
N <- 3 # Number of individuals
times <- rep(seq(0, 30, 1/12))
rates = runif(N, .2, .5)
sc_df <- data.frame(
  id = rep(1:N, each = length(times)),
  time = rep(times, times = length(rates)),
  rate = rep(rates, each = length(times))
)
sc_df$survival <- 1 - stats::pexp(sc_df$time, sc_df$rate)

# Convert the data frame to a "survival_curves" object
as_survival_curves(sc_df)
}
\seealso{
See
}
