% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurvcure.R
\name{bernoulli}
\alias{bernoulli}
\title{Implemented link functions for the mixture cure rate model}
\usage{
bernoulli(link = "logit")
}
\arguments{
\item{link}{desired link function; currently implemented links are: logit, probit, cloglog and cauchy.}
}
\value{
A list containing the codes associated with the count distribution assumed for the latent variable N and the chosen link.
}
\description{
This function is used to specify different link functions for the count component of the mixture cure rate model.
}
