% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curves.R
\name{example_survival_curves}
\alias{example_survival_curves}
\title{Example survival curves}
\usage{
example_survival_curves(n = 1000, rates = stats::runif(n, 0.2, 0.5))
}
\arguments{
\item{n}{The number of individuals to create survival curves for.}

\item{rates}{Rate parameter for each individual. Must be a numeric vector
of length \code{n}.}
}
\description{
Create a \code{\link{survival_curves}} object for use in examples by generating survival
probabilities from from exponential survival distributions.
}
\seealso{
See \code{\link[=survival_curves]{survival_curves()}} for information about the \code{survival_curves}
class.
}
