% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurvcure.R
\name{inv_pgf}
\alias{inv_pgf}
\title{Inverse of the probability generating function}
\usage{
inv_pgf(
  formula,
  incidence = c("bernoulli", "poisson", "negbin", "bell"),
  kappa = NULL,
  zeta = NULL,
  data,
  ...
)
}
\arguments{
\item{formula}{formula specifying the linear predictor for the incidence sub-model.}

\item{incidence}{the desired incidence model.}

\item{kappa}{vector of regression coefficients associated with the incidence sub-model.}

\item{zeta}{extra negative-binomial parameter.}

\item{data}{a data.frame containing the explanatory covariates passed to the formula.}

\item{...}{further arguments passed to other methods.}
}
\value{
A vector with the values of the inverse of the desired probability generating function.
}
\description{
This function is used to specify different link functions for the count component of the promotion time cure rate model.
}
