% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurvcure.R
\name{negbin}
\alias{negbin}
\title{Implemented link functions for the promotion time cure rate model with negative binomial distribution}
\usage{
negbin(zeta = stop("'theta' must be specified"), link = "log")
}
\arguments{
\item{zeta}{The known value of the additional parameter.}

\item{link}{desired link function; currently implemented links are: log, identity and sqrt.}
}
\value{
A list containing the codes associated with the count distribution assumed for the latent variable N and the chosen link.
}
\description{
This function is used to specify different link functions for the count component of the promotion time cure rate model.
}
