% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterval.R
\name{rinterval}
\alias{rinterval}
\title{Random generation of type I and type II interval censored survival data}
\usage{
rinterval(time, tau, type = c("I", "II"), prob)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{tau}{either a vector of censoring times (for type I interval-censored survival data) or time grid of scheduled visits (for type II interval censored survival data).}

\item{type}{type of interval-censored survival data (I or II).}

\item{prob}{= 0.5 attendance probability of scheduled visit; ignored when type = I.}
}
\value{
a data.frame containing the generated random sample.
}
\description{
Function to generate a random sample of type I and type II interval censored survival data.
}
