% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv-package.R
\name{rsurv-package}
\alias{rsurv-package}
\alias{rsurv}
\title{The 'rsurv' package.}
\description{
Random generation of survival data based on different survival regression models available in the literature, including Accelerated Failure Time (AFT) model, Proportional Hazard (PH) model, Proportional Odds (PO) model and the Yang & Prentice (YP) model.

_PACKAGE
}
\references{
\insertRef{2021_Demarqui}{rsurv}

\insertRef{2005_YangPrentice}{rsurv}
}
