% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv}
\alias{rsurv}
\alias{rsurv.survival_curves}
\title{Random generation from survival curve}
\usage{
rsurv(object, ...)

\method{rsurv}{survival_curves}(object, n_rep = 1, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}

\item{n_rep}{The number of times to randomly draw survival times per
individual (\code{id}) in \code{object}.}
}
\value{
A vector with randomly sampled survival times for individuals
(\code{id}s) in \code{object} repeated \code{n_rep} times. The returned vector is of length
\code{length(unique(object$id))} times \code{n_rep}.
}
\description{
\code{rsurv()} is a generic function for randomly drawing survival times for multiple
individuals very efficiently. The most flexible function is
\code{rsurv.survival_curves()}, which will randomly draw survival times from
survival curves.
}
\details{
Survival times are randomly sampled from survival curves
by partitioning times into intervals and using the survival probabilities
to compute the probability of dying within each interval, conditional on
survival up until that interval. Death is simulated within each interval from
a Bernoulli distribution and a given individual's observed death time is
the time at the end of the first interval for which that individual is
simulated to have died. This process is repeated across all individuals.
The code is written in \verb{C++} so looping across individuals is fast.
}
\examples{
# Survival curves for multiple subjects based on exponential distributions
set.seed(21)
N <- 1000 # 1000 individuals
rates <- runif(N, .2, .5)
sc <- example_survival_curves(n = N, rates = rates)
  
# Random number generation of 1000 individuals from (1) arbitrary survival
# curves and (2) exponential distributions
sim_rsurv <- rsurv(sc)
sim_rexp <- rexp(N, rate = rates)

summary(sim_rsurv)
summary(sim_rexp)

# Random number generation of 1000 individuals replicated 2 times
sim_rsurv <- rsurv(sc, n_rep = 2)
summary(sim_rsurv)
}
\seealso{
A \code{survival_curves} object can be constructed using the function
\code{\link[=survival_curves]{survival_curves()}} or converted from an existing data frame using
\code{\link[=as_survival_curves.data.frame]{as_survival_curves.data.frame()}}.
}
