% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rypreg.R
\name{rypreg}
\alias{rypreg}
\title{Random generation from Yang and Prentice models}
\usage{
rypreg(u, formula, baseline, beta, phi, dist = NULL, data, ...)
}
\arguments{
\item{u}{a numeric vector of quantiles.}

\item{formula}{formula specifying the linear predictors.}

\item{baseline}{the name of the baseline survival distribution.}

\item{beta}{vector of short-term regression coefficients.}

\item{phi}{vector of long-term regression coefficients.}

\item{dist}{an alternative way to specify the baseline survival distribution.}

\item{data}{data frame containing the covariates used to generate the survival times.}

\item{...}{further arguments passed to other methods.}
}
\value{
a numeric vector containing the generated random sample.
}
\description{
Function to generate a random sample of survival data from Yang and Prentice models.
}
\examples{
\donttest{
library(rsurv)
n <-  1000
simdata <- data.frame(
  age = rnorm(n),
  sex = sample(c("f", "m"), size = n, replace = TRUE)
) \%>\%
  mutate(
    t = rypreg(runif(n), ~ age+sex, beta = c(1, 2), phi = c(-1, 2),
                dist = "weibull", shape = 1.5, scale = 1),
    c = runif(n, 0, 10)
  ) \%>\%
  rowwise() \%>\%
  mutate(
    time = min(t, c),
    status = as.numeric(time == t)
  )
glimpse(simdata)
}

}
