% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curves.R
\name{survival_curves}
\alias{survival_curves}
\title{Survival curves}
\usage{
survival_curves(time, id, survival, sort = TRUE)
}
\arguments{
\item{time}{The times at which the survival probabilities were computed at.}

\item{id}{An identifier for each individual. May either be a numeric vector,
integer vector, character vector, or factor.}

\item{survival}{A vector of estimates of the survival probabilities.}

\item{sort}{Whether to sort the returned data frame to ensure that
rows are sorted by \code{id} and\code{time} as required by \code{\link[=rsurv.survival_curves]{rsurv.survival_curves()}}.}
}
\description{
Create a survival curves object, which stores survival probabilities at
various time points for multiple individuals.
}
\examples{
# Compute survival probabilities by individual and time
id <- 1:3 # 3 individuals
n_id <- length(id)
times1 <- seq(0, 30, 1/12) # Times for 1 individual
n_times <- length(times1)
times <- rep(times1, length(id)) # Times for all individuals
rates <- rep(runif(n_id, .2, .5), each = n_times) 
surv <- 1 - pexp(times, rates) # Survival probabilities

# Construct a survival curves object
sc <- survival_curves(time = times, id = rep(id, each = n_times),
                      survival = surv)
class(sc)
head(sc)
tail(sc)
                
}
\seealso{
\code{survival_curves} objects can can also be constructed from
a \code{\link{data.frame}} using \code{\link[=as_survival_curves]{as_survival_curves()}}. The purpose of the
\code{survival_curves} class is for storing survival curves in a standardized
format so that survival times can be randomly drawn from them using
\code{\link[=rsurv.survival_curves]{rsurv.survival_curves()}}.
}
